feat: Add email, role, age, and gender fields to user registration
The code changes in `AuthController.php` add the functionality to include the email, role, age, and gender fields in the user registration process. The `register` method now accepts these fields from the request and assigns them to the `$data` array. Additionally, a new private method `CreateName` is added to generate a username based on the email and role.

Recent user commits:
- chore: Update database name to SerenityDB
- chore: Add validation rules for age, gender, and role in registration form
- add models to insert information in the database
- update tables of database

Recent repository commits:
- chore: Update database name to SerenityDB
- chore: Add validation rules for age, gender, and role in registration form
- add models to insert information in the database
- update tables of database